Move utility methods outside of ignorelist package
